requirements.txt
PyYaml
hdbcli
